+++ python/ray/rllib/dqn/__init__.py
@@ -1,3 +1,3 @@
-from ray.rllib.dqn.simple import DQN, DEFAULT_CONFIG
+from ray.rllib.dqn.dqn import DQN, DEFAULT_CONFIG
 
 __all__ = ["DQN", "DEFAULT_CONFIG"]
